% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean up names from EpiProfile.}
\usage{
clean_names(df, label_col, ...)
}
\arguments{
\item{df}{A dataframe containing histone mass spectrometry data.}

\item{label_col}{A column containing PTM/fragment names.}

\item{...}{Optional.}
}
\value{
Two dataframes with adjusted labels, up to 4 co-occurring PTMs per fragment.
}
\description{
\code{clean_names} takes a dataframe with histone mass spectrometry data and outputs two non-transformed dataframes with user-friendly PTM names.
}
\examples{
\dontrun{
clean_names(df = THP1_histones, label_col = "PTM")
}
}
