% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histone_epi.R
\name{histone_epi}
\alias{histone_epi}
\title{Process individual data for histone proteome.}
\usage{
histone_epi(df, label_col, base_cols, target_cols, ...)
}
\arguments{
\item{df}{A dataframe containing histone mass spectrometry data. This dataframe should include measurements for two conditions.}

\item{label_col}{A column containing PTM/fragment names.}

\item{base_cols}{A list of columns containing baseline values.}

\item{target_cols}{A list of columns containing average values for your condition of interest.}

\item{...}{Optional.}
}
\value{
A dataframe with labels, log2(FC), T scores, and P-values
}
\description{
\code{histone_epi} takes a dataframe with histone mass spectrometry data and outputs a log2-transformed fold change as well as P-values and a T score.
}
\examples{
\dontrun{
histone_epi(df = THP1_single_raw, label_col = "PTM",
base_cols = c("C1", "C2", "C3", "C4", "C5"),
target_cols = c("I1", "I2", "I3", "I4", "I5"))
}
}
