% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histone_summary.R
\name{histone_summary}
\alias{histone_summary}
\title{Process summary data for histone proteome.}
\usage{
histone_summary(
  df,
  label_col,
  base_cols,
  base_std = NA,
  target_col,
  target_std,
  ...
)
}
\arguments{
\item{df}{A dataframe containing histone mass spectrometry data. This dataframe should include summary statistics (average value, standard deviation) for at least two conditions.}

\item{label_col}{A column containing PTM/fragment names.}

\item{base_cols}{A list of at least one column from which to generate a baseline.}

\item{base_std}{If more than one column is supplied to \code{base_cols}, leave this as NA. Otherwise, if only one column is supplied, provide the standard deviation for that column.}

\item{target_col}{A column containing average values for your condition of interest.}

\item{target_std}{A column containing standard deviations for your condition of interest.}

\item{...}{Optional.}
}
\value{
A dataframe with labels, log2(FC), T scores, and P-values
}
\description{
\code{histone_summary} takes a dataframe with histone mass spectrometry data and outputs a log2-transformed fold change as well as P-values and a T score.
}
\examples{
\dontrun{
histone_summary(df = MCF7_histones, label_col = "PTM",
base_cols = c("293 ave (5 reps)", "HaCAT ave (5 reps)", "hESC ave (6 reps)",
"HFF ave (7 reps)", "Mdm13 ave (5 reps)"), base_std = NA,
target_col = "MCF7 ave (5 reps)", target_std = "MCF7 std")
}
}
