% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCF7_histones.R
\docType{data}
\name{MCF7_histones}
\alias{MCF7_histones}
\title{MCF-7 histone proteome dataset.}
\format{
A data frame with 37 rows and 12 variables:
\describe{
\item{PTM}{PTM descriptions.}
\item{MCF7 std}{Standard deviation in MCF-7 cells.}
\item{MCF7 ave (5 reps)}{Average expression in MCF-7 cells.}
\item{293 ave (5 reps)}{Average expression in HEK293 cells.}
\item{HaCAT ave (5 reps)}{Average expression in HaCAT cells.}
\item{hESC ave (6 reps)}{Average expression in hESC cells.}
\item{HFF ave (7 reps)}{Average expression in HFF cells.}
\item{Mdm13 ave (5 reps)}{Average expression in Mdm13 cells.}
\item{ctrl ave}{Average expression in control cell types (HEK293, HaCAT, hESC, HFF, Mdm13).}
\item{log2(FC)}{The log2 transformed fold change in expression between MCF-7 and control cells.}
\item{T score}{The calculated T score using standard deviation.}
\item{P value}{The corresponding P value.}
}
}
\source{
\url{https://doi.org/10.1186/1756-8935-6-20}
}
\usage{
MCF7_histones
}
\description{
Histone proteome data from Garcia et al. (2013).
}
\keyword{datasets}
