% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCF7_total.R
\docType{data}
\name{MCF7_total}
\alias{MCF7_total}
\title{MCF-7 total proteome dataset.}
\format{
A data frame with 8,317 rows and 6 variables:
\describe{
\item{Accession}{UniProt accession numbers.}
\item{M90/F Ratio}{Raw ratio for fold change calculation.}
\item{log2(FC)}{The log2 transformed fold change of M90/F cells.}
\item{P value}{Calculated unadjusted P value.}
\item{ENTREZID}{ENTREZ ID for pathway analysis.}
\item{SYMBOL2}{Common gene symbols.}
}
}
\source{
\url{https://doi.org/10.1021/acs.jproteome.4c00563}
}
\usage{
MCF7_total
}
\description{
Total proteome data from Yang et al. (2024). M90 condition: MCF-7 cells grown in 90 Pa matrix. F condition (control): MCF-7 cells grown on flat culture surface.
}
\keyword{datasets}
