% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_net.R
\name{multi_net}
\alias{multi_net}
\title{Generate a multi-level proteome network from total proteome, phosphoproteome, and epiproteome data.}
\usage{
multi_net(
  genes_df,
  total_df,
  phospho_df,
  symbol_col,
  fc_col,
  epi_single_df,
  epi_symbol_col,
  epi_fc_col,
  ...
)
}
\arguments{
\item{genes_df}{A dataframe containing the exported information from EpiFactors.}

\item{total_df}{A dataframe containing total proteome data.}

\item{phospho_df}{A dataframe containing phosphoproteome data.}

\item{symbol_col}{A column in both \code{total_df} and \code{phospho_df} containing the gene symbol names for identified proteins.}

\item{fc_col}{A column in both \code{total_df} and \code{phospho_df} containing the log2-transformed fold change values.}

\item{epi_single_df}{A dataframe containing histone proteome data. If using the full pipeline, make sure this is the single PTM dataframe and not fragments.}

\item{epi_symbol_col}{A column in \code{epi_single_df} containing the gene symbol names for identified PTMs.}

\item{epi_fc_col}{A column in \code{epi_single_df} containing the log2-transformed fold change values for identified PTMs.}

\item{...}{Optional.}
}
\value{
An object with the following attributes:
\itemize{
\item \code{grid}: A plot displaying the complexes, PTMs, and networks in an intersecting grid.
\item \code{complex_list}: A list of the complexes present in the network.
\item \code{member_list}: A list of the complexes, their complex members (proteins), and associated PTMs present in the network.
\item \code{PTM_list}: A list of the identified PTMs and their differential expression.
\item \code{gene_list}: A subset of the \code{genes_df} variable containing proteins identified.
\item \code{net_edges}: The edges connecting network vertices, which identify member-PTM relationships.
\item \code{net_vertices}: The vertices (PTMs and members) in the network.
\item \code{ts}: A larger set of information on the proteins in the network. This contains information including statistical significance values, fold change values, and proteome source identification.
}
}
\description{
\code{multi_net} generates an R object ready to generate a network in Python given three datasets.
}
\examples{
\dontrun{
multi_net()
}
}
