% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THP1_phospho.R
\docType{data}
\name{THP1_phospho}
\alias{THP1_phospho}
\title{THP-1 phosphoproteome dataset.}
\format{
A data frame with 6,121 rows and 4 variables:
\describe{
\item{log2(FC)}{The log2 transformed fold change of infected/control cells.}
\item{P value}{Calculated unadjusted P value.}
\item{ENTREZID}{ENTREZ ID for pathway analysis.}
\item{SYMBOL2}{Common gene symbols.}
\item{UNIPROT}{UNIPROT IDs for the identified proteins.}
}
}
\source{
\url{https://doi.org/10.1021/acs.jproteome.9b00895}
}
\usage{
THP1_phospho
}
\description{
Phosphoproteome data from Agarwal et al. (2021). Infected condition: THP-1 macrophages exposed to M. bovis BCG for 48 hours at a multiplicity of 10.
}
\keyword{datasets}
