% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_epi.R
\name{volcano_epi}
\alias{volcano_epi}
\title{Generate a volcano plot for epiproteome data.}
\usage{
volcano_epi(df, fc_col, p_col, fc_thr = 0.5, p_thr = 0.05, label_col = NA, ...)
}
\arguments{
\item{df}{A data frame containing points to be plotted.}

\item{fc_col}{A column in \code{df} with fold change values to be plotted.}

\item{p_col}{A column in \code{df} with P-values to be plotted. These values should not be log-transformed.}

\item{fc_thr}{A scalar value indicating the fold change threshold. Default = 0.5.}

\item{p_thr}{A scalar value indicating the P-value threshold. Default = 0.05.}

\item{label_col}{A column in \code{df} with which points should be labeled. If \code{label_col = NA}, default labels will be applied to significant points.}

\item{...}{Optional.}
}
\value{
A ggplot2 object.
}
\description{
\code{volcano_epi} generates a volcano plot given data and threshold values.
}
\examples{
\dontrun{
volcano_epi(df = total_raw, fc_col = "FC", p_col = "Pval", label_col = "SYMBOL2") +
ggtitle("Volcano plot") + xlab("log2FC") + ylab("-log10P")
}
}
