% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_groups.R
\name{volcano_groups}
\alias{volcano_groups}
\title{Generate a volcano plot with groups of interest.}
\usage{
volcano_groups(
  df,
  fc_col,
  p_col,
  fc_thr = 0.5,
  p_thr = 0.05,
  label_col = NA,
  compute_fc_log2 = FALSE,
  group_list = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame containing points to be plotted.}

\item{fc_col}{A column in \code{df} with fold change values to be plotted.}

\item{p_col}{A column in \code{df} with P-values to be plotted. These values should not be log-transformed.}

\item{fc_thr}{A scalar value indicating the fold change threshold. Default = 0.5.}

\item{p_thr}{A scalar value indicating the P-value threshold. Default = 0.05.}

\item{label_col}{A column in \code{df} with values to be used if groups are not provided.}

\item{compute_fc_log2}{A Boolean to determine if the \code{fc_col} values should be additionally log2-transformed. Default = FALSE.}

\item{group_list}{A named list of lists where each list is a group of peptides, labeled with the group name.}

\item{...}{Optional.}
}
\value{
A ggplot2 object.
}
\description{
\code{volcano_groups} generates a volcano plot given data, threshold values, and a subset of points to label.
}
\examples{
\dontrun{
volcano_groups(df = THP1_total, fc_col = "log2(FC)", p_col = "P value",
  label_col = "SYMBOL2", group_list = list("Group 1" =
   c("MORF4L1", "SAMSN1", "DTX3L", "MBD3",  "HSF1",
  "EYA3", "NCAPD2", "TAF9", "PPP4C", "ZNF638"),
  "Group 2" = c("CD9", "CD82", "NFKB1", "BST2", "CTSH"))) +
  ggtitle("Grouped volcano plot") +
  xlab("log2FC") + ylab("-log10P")
}
}
