% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_groups.R, R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Generate a volcano plot with groups of interest.}
\usage{
volcanoplot(
  df,
  fc_col,
  p_col,
  fc_thr = 0.5,
  p_thr = 0.05,
  label_col = NA,
  compute_fc_log2 = FALSE,
  ...
)

volcanoplot(
  df,
  fc_col,
  p_col,
  fc_thr = 0.5,
  p_thr = 0.05,
  label_col = NA,
  compute_fc_log2 = FALSE,
  ...
)
}
\arguments{
\item{df}{A data frame containing points to be plotted.}

\item{fc_col}{A column in \code{df} with fold change values to be plotted.}

\item{p_col}{A column in \code{df} with P-values to be plotted. These values should not be log-transformed.}

\item{fc_thr}{A scalar value indicating the fold change threshold. Default = 0.5.}

\item{p_thr}{A scalar value indicating the P-value threshold. Default = 0.05.}

\item{label_col}{A column in \code{df} with which points should be labeled. If \code{label_col = NA}, default labels will be applied to significant points.}

\item{compute_fc_log2}{A Boolean to determine if the \code{fc_col} values should be additionally log2-transformed. Default = FALSE.}

\item{...}{Optional.}

\item{groups}{A list of named lists where each list is a group of peptides, labeled with the group name.}
}
\value{
A ggplot2 object.

A ggplot2 object.
}
\description{
\code{volcano_groups} generates a volcano plot given data, threshold values, and a subset of points to label.

\code{volcanoplot} generates a volcano plot given data, threshold values, and a subset of points to label.
}
\examples{
\dontrun{
volcano_groups(df = THP1_total, fc_col = "log2(FC)", p_col = "P value", label_col = "SYMBOL2") + ggtitle("Grouped volcano plot") + xlab("log2FC") + ylab("-log10P")
}
\dontrun{
volcanoplot(df = total_raw, fc_col = "FC", p_col = "Pval", label_col = "SYMBOL2") + ggtitle("Volcano plot") + xlab("log2FC") + ylab("-log10P")
}
}
